% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnegbin.R
\name{dnegbin}
\alias{dnegbin}
\alias{dnegbin.alphabeta}
\alias{rnegbin.alphabeta}
\title{The negative binomial distribution}
\usage{
dnegbin(k, r, p)

dnegbin.alphabeta(k, alpha, beta)

rnegbin.alphabeta(n, alpha, beta)
}
\arguments{
\item{k}{numeric. Vector of observed counts.}

\item{r}{scalar, positive.}

\item{p}{scalar, probability.}

\item{alpha}{scalar, positive.}

\item{beta}{scalar, positive.}

\item{n}{scalar, positive. Number of samples.}
}
\description{
rnegbin.alphabeta simulates data from a negative binomial distribution and dnegbin and dnegbin.alphabeta compute the density.
}
\details{
The functions with \code{alphabeta} use the parameterization with alpha and beta, while \code{dnegbin} uses the parameterization with r and p.

The functions \code{rgamma} and \code{rstats} from package stats are used for the simulation of negative binomial data.

dnegbin and dnegbin.alphabeta can handle non-integer counts (without warning), because the binomial coefficient is implemented with a gamma function.
}
\examples{

# see rnegbinmix

}
\author{
Johanna Bertl
}
