% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/EM.R
\name{EM}
\alias{EM}
\title{EM algorithm for the negative binomial mixture model}
\usage{
EM(theta.null, x.syn, x.non, c1, c2, iter, ui, ci)
}
\arguments{
\item{theta.null}{}

\item{x.syn}{}

\item{x.non}{}

\item{c1}{}

\item{c2}{}

\item{iter}{}

\item{ui}{}

\item{ci}{}
}
\description{
EM algorithm to estimate the parameters of the negative binomial mixture model. Currently only for mixtures of 2 negbinoms as defined in \link{negbinmix::rnegbinmix}.
}
\details{
In the M step, contrained optimization is conducted with the constrOptim function. ui and ci are defined accordingly (see \link{constrOptim}.
}
\author{
Johanna Bertl
}

