% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Q.R
\name{Q}
\alias{Q}
\title{The function Q for the negative binomial mixture model}
\usage{
Q(theta, theta.prime, x.syn, x.non, c1, c2)
}
\description{
Currently only for mixtures of 2 negbinoms.
}
\author{
Johanna Bertl
}

