% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Q_proportions.R
\name{Q_proportions}
\alias{Q_proportions}
\title{The function Q for the EM algorithm to estimate the parameters p1, ..., pk-1 in the negative binomial mixture model}
\usage{
Q_proportions(theta, theta.prime, x.non, alpha, beta, cvec)
}
\arguments{
\item{theta}{c(p1, p2, ..., pk-1)}

\item{theta.prime}{c(p1.prime, p2.prime, ..., p(k-1).prime)}

\item{x.non}{vector of non-synonymous mutation count}

\item{alpha}{}

\item{beta}{}

\item{cvec}{vector of k positive values c(c1, ..., ck)}

\item{x.syn}{vector of synonymous mutation count}
}
\description{
This function is used by the function EM_proportions. The parameters alpha and beta are assumed to be known. The function negQ_proportions = -Q_proportions can be used for minimization.
}
\author{
Johanna Bertl
}

