% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Q_Non.R
\name{QNon}
\alias{QNon}
\title{The function Q for the negative binomial mixture model of the non-synonymous mutations only!}
\usage{
QNon(theta, theta.prime, x.non, c1, c2)
}
\arguments{
\item{theta}{c(alpha, beta, p1)}

\item{theta.prime}{c(alpha.prime, beta.prime, p1.prime)}

\item{x.non}{vector of non-synonymous mutation count}

\item{c1}{positive scalar}

\item{c2}{positive scalar}
}
\description{
Currently only for mixtures of 2 negbinoms.
}
\details{
The function negQ is -Q (for minimization).

The function Q_fixed_prop is for the function EM_fixed_prop, where the mixture proportions are assumed fixed. Here, theta = c(alpha, beta).
}
\author{
Johanna Bertl
}

