% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rnegbinmix.R
\name{rnegbinmix}
\alias{rnegbinmix}
\title{Simulating data from a negative binomial mixture distribution}
\usage{
rnegbinmix(n, alpha, beta, c, p)
}
\arguments{
\item{n}{}

\item{alpha,}{scalar, positive}

\item{beta}{positive}

\item{c}{numeric, positive, length k}

\item{p}{numeric, probability vector, length k}
}
\description{
rnegbinmix simulates data from a negative binomial mixture distribution.
}
\details{
The negative binomial mixture is defined as follows:

The function \code{rnegbin} from package MASS is used for the simulation of negative binomial data.
}
\author{
Johanna Bertl
}

